pyproject.toml
[tool.poetry]
name = "stock-ai-investor"
version = "0.1.0"
description = "LLM-powered AI agent for stock market investment."
authors = ["Grigorios Maragkakis <grigorios.maragkakis@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
jinja2 = "^3.1.5"
requests = "^2.32.3"
yfinance = "^0.2.54"
pandas = "^2.2.3"
numpy = "^2.2.3"
openai = "^1.64.0"
pydantic-settings = "^2.8.0"
python-dotenv = "^1.0.1"

[poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"